requirements.txt
streamlit>=1.44,<2.0
pdfplumber>=0.11,<1.0
pypdf>=5.0,<6.0
pytesseract>=0.3.10,<1.0
Pillow>=10.4,<12.0
transformers>=4.45,<5.0
torch>=2.4,<3.0
google-genai>=1.12,<2.0
PyYAML>=6.0,<7.0
pydantic>=2.8,<3.0
pytest>=8.2,<9.0
reportlab>=4.2,<5.0
